+++ HSTracker/Core/Settings.swift
@@ -255,6 +255,8 @@ final class Settings {
     static var showMulliganToast: Bool
     @UserDefault(key: Settings.show_flavor_text, defaultValue: true)
     static var showFlavorText: Bool
+    @UserDefault(key: Settings.show_board_entry_order, defaultValue: false)
+    static var showBoardEntryOrder: Bool
     @UserDefault(key: Settings.enable_mulligan_guide, defaultValue: true)
     static var enableMulliganGuide: Bool
     @UserDefault(key: Settings.enable_mulligan_gv2, defaultValue: true)
@@ -908,6 +910,7 @@ extension Settings {
     static let show_experience_counter = "show_experience_counter"
     static let show_mulligan_toast = "show_mulligan_toast"
     static let show_flavor_text = "show_flavor_text"
+    static let show_board_entry_order = "show_board_entry_order"
     static let enable_mulligan_guide = "enable_mulligan_guide"
     static let enable_mulligan_gv2 = "enable_mulligan_gv2"
     static let show_mulligan_guide_pre_lobby = "show_mulligan_guide_pre_lobby"