+++ HSTracker/UIs/Preferences/TrackersPreferences.swift
@@ -34,6 +34,7 @@ class TrackersPreferences: PreferencePaneController, PreferencePane {
     @IBOutlet var showExperienceCounter: NSButton!
     @IBOutlet var showMulliganToast: NSButton!
     @IBOutlet var showFlavorText: NSButton!
+    @IBOutlet var showBoardEntryOrder: NSButton!
     @IBOutlet var enableMulliganGuide: NSButton!
     @IBOutlet var enableMulliganGV2: NSButton!
     @IBOutlet var showMulliganGuidePreLobby: NSButton!
@@ -70,6 +71,7 @@ class TrackersPreferences: PreferencePaneController, PreferencePane {
         showExperienceCounter.state = Settings.showExperienceCounter ? .on : .off
         showMulliganToast.state = Settings.showMulliganToast ? .on : .off
         showFlavorText.state = Settings.showFlavorText ? .on : .off
+        showBoardEntryOrder.state = Settings.showBoardEntryOrder ? .on : .off
 
         theme.selectItem(at: themes.firstIndex(of: Settings.theme) ?? 0)
         allowFullscreen.state = Settings.canJoinFullscreen ? .on : .off
@@ -151,6 +153,8 @@ class TrackersPreferences: PreferencePaneController, PreferencePane {
             Settings.showMulliganToast = showMulliganToast.state == .on
         } else if sender == showFlavorText {
             Settings.showFlavorText = showFlavorText.state == .on
+        } else if sender == showBoardEntryOrder {
+            Settings.showBoardEntryOrder = showBoardEntryOrder.state == .on
         } else if sender == enableMulliganGuide {
             Settings.enableMulliganGuide = enableMulliganGuide.state == .on
             let game = AppDelegate.instance().coreManager.game